Moved toast function to LogContext, started handleReadLayer

+++ src/js/main/modules/util.tsx
@@ -1,4 +1,3 @@
-import { ToastQueue } from "@react-spectrum/toast";
 import { ProfileKey } from "../modules";
 
 export const toastTimeout: number = 1500;
@@ -12,40 +11,12 @@ export const menuIconMargin: string = "size-75";
 export const menuTextMargin: string = "size-250";
 export const iconMarginAdjust: number = 6.5;
 
-// All Functions with Try Catch Blocks have optional hideSuccess
-// parameter to suppress success messages, default is true
-
 export function logCaughtError(caughtError: TypeError, customMessage?: string) {
   let message = "--- CAUGHT ERROR ---\n\n";
   if (customMessage) {
     message += customMessage + "\n\n";
   }
   console.log(message + caughtError.message);
-}
-
-export function postToast(
-  style: "positive" | "negative" | "info" | "neutral",
-  message: string,
-  appLog?: (message: string, header?: string) => void
-) {
-  if (appLog) {
-    appLog(message, style.toUpperCase());
-  }
-
-  switch (style) {
-    case "positive":
-      ToastQueue.positive(message, { timeout: toastTimeout });
-      break;
-    case "negative":
-      ToastQueue.negative(message, { timeout: toastTimeoutExtended });
-      break;
-    case "info":
-      ToastQueue.info(message, { timeout: toastTimeout });
-      break;
-    default:
-      ToastQueue.neutral(message, { timeout: toastTimeout });
-      break;
-  }
 }
 
 export type UnitItem = {
@@ -68,27 +39,41 @@ export function getLocalStorageLength(): number {
   return localStorage.length || 0;
 }
 
-export function getLocalStorageList(showSuccess = false): string[] {
+export function getLocalStorageList(
+  postToast?: (
+    style: "positive" | "negative" | "info" | "neutral",
+    message: string
+  ) => void,
+  showSuccess = false
+): string[] {
   let fileList: string[] = [];
   try {
     for (let i = 0; i < getLocalStorageLength(); i++) {
       fileList.push(localStorage.key(i) || "");
     }
-    if (showSuccess) {
+    if (showSuccess && postToast) {
       postToast("positive", "Loaded File List: " + fileList.length + fileList);
     }
   } catch (error) {
-    error instanceof Error
-      ? postToast("negative", error.message)
-      : postToast(
-          "negative",
-          "Something went wrong trying to read local storage"
-        );
+    if (postToast) {
+      error instanceof Error
+        ? postToast("negative", error.message)
+        : postToast(
+            "negative",
+            "Something went wrong trying to read local storage"
+          );
+    }
   }
   return fileList;
 }
 
-export function getLocalStorageProfiles(showSuccess = false): string[] {
+export function getLocalStorageProfiles(
+  postToast?: (
+    style: "positive" | "negative" | "info" | "neutral",
+    message: string
+  ) => void,
+  showSuccess = false
+): string[] {
   let fileList: string[] = [];
   let fileCount = getLocalStorageLength();
   try {
@@ -98,19 +83,21 @@ export function getLocalStorageProfiles(showSuccess = false): string[] {
         fileList.push(key);
       }
     }
-    if (showSuccess) {
+    if (showSuccess && postToast) {
       postToast(
         "positive",
         "Loaded Profile List: " + fileList.length + fileList
       );
     }
   } catch (error) {
-    error instanceof Error
-      ? postToast("negative", error.message)
-      : postToast(
-          "negative",
-          "Something went wrong trying to load stored profiles"
-        );
+    if (postToast) {
+      error instanceof Error
+        ? postToast("negative", error.message)
+        : postToast(
+            "negative",
+            "Something went wrong trying to load stored profiles"
+          );
+    }
   }
 
   return fileList;
@@ -119,35 +106,52 @@ export function getLocalStorageProfiles(showSuccess = false): string[] {
 export function writeLocalStorage(
   key: string,
   value: any,
+
+  postToast?: (
+    style: "positive" | "negative" | "info" | "neutral",
+    message: string
+  ) => void,
   showSuccess = false
 ) {
   try {
     localStorage.setItem(key, JSON.stringify(value));
-    if (showSuccess) {
+    if (showSuccess && postToast) {
       postToast("positive", `Saved ${key} successfully`);
     }
   } catch (error) {
-    if (error instanceof Error) {
-      postToast("negative", error.message);
-    } else {
-      postToast("negative", "Something went wrong trying to save");
+    if (postToast) {
+      if (error instanceof Error) {
+        postToast("negative", error.message);
+      } else {
+        postToast("negative", "Something went wrong trying to save");
+      }
     }
   }
 }
 
-export function readLocalStorage(key: string, showSuccess = false) {
+export function readLocalStorage(
+  key: string,
+
+  postToast?: (
+    style: "positive" | "negative" | "info" | "neutral",
+    message: string
+  ) => void,
+  showSuccess = false
+) {
   var storedData;
   try {
     storedData = localStorage.getItem(key);
   } catch (error) {
-    if (error instanceof Error) {
-      postToast("negative", error.message);
-    } else {
-      postToast("negative", "Something went wrong trying to load");
+    if (postToast) {
+      if (error instanceof Error) {
+        postToast("negative", error.message);
+      } else {
+        postToast("negative", "Something went wrong trying to load");
+      }
     }
   }
   if (storedData) {
-    if (showSuccess) {
+    if (showSuccess && postToast) {
       postToast("info", "Loaded successfully");
     }
 
@@ -157,24 +161,37 @@ export function readLocalStorage(key: string, showSuccess = false) {
   }
 }
 
-export function deleteLocalStorage(key: string, showSuccess = false) {
+export function deleteLocalStorage(
+  key: string,
+
+  postToast?: (
+    style: "positive" | "negative" | "info" | "neutral",
+    message: string
+  ) => void,
+  showSuccess = false
+) {
   try {
     localStorage.removeItem(key);
-    if (showSuccess) {
+    if (showSuccess && postToast) {
       postToast("positive", `Deleted ${key} successfully`);
     }
   } catch (error) {
-    if (error instanceof Error) {
-      postToast("negative", error.message);
-    } else {
-      postToast(
-        "negative",
-        "Something went wrong trying to delete from storage"
-      );
+    if (postToast) {
+      if (error instanceof Error) {
+        postToast("negative", error.message);
+      } else {
+        postToast(
+          "negative",
+          "Something went wrong trying to delete from storage"
+        );
+      }
     }
   }
 }
 
 export function formatFieldName(fieldName: string): string {
-  return fieldName.toLowerCase().replace(/\s+/g, "_");
+  return fieldName
+    .toLowerCase()
+    .replace(/\s+/g, "_")
+    .replace(/[^a-zA-Z0-9]/g, "");
 }
